Added a new search field and updated the product forms

+++ db/migrate/20160605202000_add_price_to_products.rb
@@ -1,5 +1,5 @@
 class AddPriceToProducts < ActiveRecord::Migration
   def change
-    add_column :products, :price, :integer
+    add_column :products, :price, :decimal
   end
 end
